+++ landingzones/tests/test_views_ajax_taskflow.py
@@ -11,7 +11,7 @@ from django.urls import reverse
 
 # Projectroles dependency
 from projectroles.app_settings import AppSettingAPI
-from projectroles.models import SODAR_CONSTANTS
+from projectroles.models import SODARUser, SODAR_CONSTANTS
 
 # Samplesheets dependency
 from samplesheets.tests.test_io import SampleSheetIOMixin
@@ -55,6 +55,13 @@ class TestZoneIrodsListRetrieveAjaxView(
     TaskflowViewTestBase,
 ):
     """Tests for ZoneIrodsListRetrieveAjaxView with iRODS and taskflow"""
+
+    def _assert_coll_setting(self, user: SODARUser, expected: bool):
+        """Assert collection including setting value for user"""
+        self.assertEqual(
+            app_settings.get(APP_NAME, 'zone_file_list_colls', user=user),
+            expected,
+        )
 
     def setUp(self):
         super().setUp()
@@ -105,6 +112,7 @@ class TestZoneIrodsListRetrieveAjaxView(
                 }
                 for i in range(len(ZONE_COLLS))
             ],
+            'include_colls': True,
             'count': 3,
             'page': 1,
             'page_count': 1,
@@ -112,6 +120,57 @@ class TestZoneIrodsListRetrieveAjaxView(
             'previous': '',
         }
         self.assertEqual(response.data, expected)
+
+    def test_get_colls_true(self):
+        """Test GET with colls explicitly set True"""
+        # Set user app setting to false (default=True)
+        app_settings.set(
+            APP_NAME, 'zone_file_list_colls', False, user=self.user_owner
+        )
+        self._assert_coll_setting(self.user_owner, False)
+        self.make_zone_taskflow(self.zone, colls=ZONE_COLLS)
+        with self.login(self.user_owner):
+            response = self.client.get(self.url + '?page=1&colls=1')
+        self.assertEqual(response.status_code, 200)
+        expected = {
+            'results': [
+                {
+                    'name': ZONE_COLLS[i],
+                    'type': IRODS_TYPE_COLL,
+                    'path': iRODSPath(self.zone_path, ZONE_COLLS[i]),
+                }
+                for i in range(len(ZONE_COLLS))
+            ],
+            'include_colls': True,
+            'count': 3,
+            'page': 1,
+            'page_count': 1,
+            'next': '',
+            'previous': '',
+        }
+        self.assertEqual(response.data, expected)
+        # User app setting should be updated
+        self._assert_coll_setting(self.user_owner, True)
+
+    def test_get_colls_false(self):
+        """Test GET with colls=False"""
+        self._assert_coll_setting(self.user_owner, True)
+        self.make_zone_taskflow(self.zone, colls=ZONE_COLLS)
+        with self.login(self.user_owner):
+            response = self.client.get(self.url + '?page=1&colls=0')
+        self.assertEqual(response.status_code, 200)
+        expected = {
+            'results': [],
+            'include_colls': False,
+            'count': 0,
+            'page': 1,
+            'page_count': 0,
+            'next': '',
+            'previous': '',
+        }
+        self.assertEqual(response.data, expected)
+        # Setting should be set False
+        self._assert_coll_setting(self.user_owner, False)
 
     def test_get_file(self):
         """Test GET with file in iRODS"""
@@ -135,6 +194,29 @@ class TestZoneIrodsListRetrieveAjaxView(
             ).strftime('%Y-%m-%d %H:%M'),
         }
         self.assertEqual(res[1], expected)
+
+    def test_get_file_colls_false(self):
+        """Test GET with file and colls=False"""
+        self.make_zone_taskflow(self.zone, colls=ZONE_COLLS)
+        coll = self.irods.collections.get(self.misc_path)
+        irods_obj = self.make_irods_object(coll, TEST_OBJ_NAME)
+        # NOTE: No .md5 file created here
+        with self.login(self.user_owner):
+            response = self.client.get(self.url + '?page=1&colls=0')
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.data['count'], 1)
+        res = response.data['results']
+        self.assertEqual(len(res), 1)
+        expected = {
+            'name': TEST_OBJ_NAME,
+            'type': IRODS_TYPE_OBJ,
+            'path': irods_obj.path,
+            'size': 1024,
+            'modify_time': irods_obj.modify_time.astimezone(
+                pytz.timezone(settings.TIME_ZONE)
+            ).strftime('%Y-%m-%d %H:%M'),
+        }
+        self.assertEqual(res[0], expected)
 
     def test_get_file_md5(self):
         """Test GET with file and corresponding MD5 file"""
@@ -168,6 +250,7 @@ class TestZoneIrodsListRetrieveAjaxView(
         self.assertEqual(response.status_code, 200)
         expected = {
             'results': [],
+            'include_colls': True,
             'count': 0,
             'page': 1,
             'page_count': 0,