+++ CHANGELOG.rst
@@ -23,6 +23,8 @@ Added
     - ``notify_alert_zone_status`` user app setting (#2205)
     - ``ZoneSettingsRetrieveAPIView`` REST API view (#2174)
     - Periodic celery task setup logging (#2275)
+    - ``zone_file_list_colls`` user app setting (#2284)
+    - File list modal collection display toggling (#2284)
 - **Samplesheets**
     - Project viewer role support (#2200)
     - Category statistics for sample count and iRODS statistics (#2213)